Finished dll

+++ doubly_linked_list/doubly_linked_list.py
@@ -52,8 +52,8 @@ class DoublyLinkedList:
     def __len__(self):
         return self.length
 
-    """Wraps the given value in a ListNode and inserts it 
-    as the new head of the list. Don't forget to handle 
+    """Wraps the given value in a ListNode and inserts it
+    as the new head of the list. Don't forget to handle
     the old head node's previous pointer accordingly."""
     def add_to_head(self, value):
         if self.head:
@@ -66,16 +66,20 @@ class DoublyLinkedList:
     """Removes the List's current head node, making the
     current head's next node the new head of the List.
     Returns the value of the removed Node."""
-    def remove_from_head(self, value):
+    def remove_from_head(self):
         if not self.head:
             return None
-        if self.head:
-            removed = self.head
+        value = self.head.value
+        if self.head is self.tail:
+            self.head = None
+            self.tail = None
+        else:
             self.delete(self.head)
-            return removed.value
+        self.length -= 1
+        return value
 
-    """Wraps the given value in a ListNode and inserts it 
-    as the new tail of the list. Don't forget to handle 
+    """Wraps the given value in a ListNode and inserts it
+    as the new tail of the list. Don't forget to handle
     the old tail node's next pointer accordingly."""
     def add_to_tail(self, value):
         new_node = ListNode(value, None, None)
@@ -89,29 +93,29 @@ class DoublyLinkedList:
         self.tail = new_node
         self.length += 1
 
-    """Removes the List's current tail node, making the 
+    """Removes the List's current tail node, making the
     current tail's previous node the new tail of the List.
     Returns the value of the removed Node."""
     def remove_from_tail(self):
         if not self.tail:
             return None
-        if self.tail.next is None:
-            tail_value = self.tail.value
+        value = self.tail.value
+        if self.head is self.tail:
             self.head = None
             self.tail = None
-            return tail_value
-        tail_value = self.tail.value
-        self.tail = self.tail.next
-        return tail_value
+        else:
+            self.delete(self.tail)
+        self.length -= 1
+        return value
 
-    """Removes the input node from its current spot in the 
+    """Removes the input node from its current spot in the
     List and inserts it as the new head node of the List."""
     def move_to_front(self, node):
         value = node.value
         self.delete(node)
         self.add_to_head(value)
 
-    """Removes the input node from its current spot in the 
+    """Removes the input node from its current spot in the
     List and inserts it as the new tail node of the List."""
     def move_to_end(self, node):
         value = node.value
@@ -120,15 +124,35 @@ class DoublyLinkedList:
 
     """Removes a node from the list and handles cases where
     the node was the head or the tail"""
-    def delete(self, node, value):
+    def delete(self, node):
+        # the list is empty -> do nothing
+        # the list is only one node
+
+        if not self.head:
+            return None
+        if self.head is self.tail:
+            self.head = None
+            self.tail = None
         if node == self.head:
             self.head = node.next
+            node.delete()
         if node == self.tail:
             self.tail = node.prev
+            node.delete()
+        else:
+            node.delete()
         self.length -= 1
-        node.delete()
 
     """Returns the highest value currently in the list"""
     def get_max(self):
-        pass
+        if not self.head:
+            return None
+        curr = self.head
+        maximum = self.head.value
+
+        while curr:
+            if curr.value > maximum:
+                maximum = curr.value
+            curr = curr.next
+        return maximum
